+++ src/renderer/component/common/form-components/form-field.jsx
@@ -6,7 +6,7 @@ import MarkdownPreview from 'component/common/markdown-preview';
 import SimpleMDE from 'react-simplemde-editor';
 import 'simplemde/dist/simplemde.min.css'; // eslint-disable-line import/no-extraneous-dependencies
 import Toggle from 'react-toggle';
-import { openEditorMenu } from 'util/contextMenu';
+import { openEditorMenu, stopContextMenu } from 'util/contextMenu';
 
 type Props = {
   name: string,
@@ -57,15 +57,10 @@ export class FormField extends React.PureComponent<Props> {
           </select>
         );
       } else if (type === 'markdown') {
-        const stopContextMenu = event => {
-          event.preventDefault();
-          event.stopPropagation();
-        };
         const handleEvents = {
-          contextmenu(codeMirror, event) {
-            openEditorMenu(event, codeMirror);
-          },
+          contextmenu: openEditorMenu,
         };
+
         input = (
           <div className="form-field--SimpleMDE" onContextMenu={stopContextMenu}>
             <SimpleMDE