+++ src/renderer/util/getMediaType.js
@@ -3,7 +3,9 @@ import mime from 'mime';
 const formats = [
   [/\.(mp4|m4v|webm|flv|f4v|ogv)$/i, 'video'],
   [/\.(mp3|m4a|aac|wav|flac|ogg|opus)$/i, 'audio'],
-  [/\.(html|htm|xml|pdf|odf|doc|docx|md|markdown|txt|epub|org)$/i, 'document'],
+  [/\.(h|go|ja|java|js|jsx|c|cpp|cs|css|rb|scss|sh|php|py)$/i, 'script'],
+  [/\.(json|csv|txt|log|md|markdown|docx|pdf|xml|yml|yaml)$/i, 'document'],
+  [/\.(pdf|odf|doc|docx|epub|org|rtf)$/i, 'e-book'],
   [/\.(stl|obj|fbx|gcode)$/i, '3D-file'],
 ];